Alteração de Int64 para int32 nos mapeamentos. Para realizar os teste no ORACLE.

+++ Hangfire.FluentNHibernateStorage/JobQueue/FetchedJob.cs
@@ -2,8 +2,8 @@
 {
     public class FetchedJob
     {
-        public long Id { get; set; }
-        public long JobId { get; set; }
+        public int Id { get; set; }
+        public int JobId { get; set; }
         public string Queue { get; set; }
     }
 }
